+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 - fixed #1 fails to format }}}
 - fixed #2 incorrect parsing of nested array 
 - fixed #7 formatting issue: nested arrays [[[]]] 
+- fixed #6 formatting issue: empty object in an array
 
 **0.00.13**
 - binary display enhancement